requirements.txt
# ── LangGraph orchestration ───────────────────────────────────────────────────
langgraph>=0.2
langchain-core>=0.3

# ── Computer Vision stack ─────────────────────────────────────────────────────
ultralytics>=8.0
sahi>=0.11
timm
torch>=2.0
torchvision
opencv-python
Pillow
numpy
matplotlib

# ── Segment Anything Model (Meta) ─────────────────────────────────────────────
# Install separately after pip install -r requirements.txt:
#   pip install git+https://github.com/facebookresearch/segment-anything.git
# segment-anything

# ── AI API clients ────────────────────────────────────────────────────────────
groq
google-genai
google-generativeai

# ── Fraud detection utilities ─────────────────────────────────────────────────
exifread          # EXIF metadata extraction (Batch 1 Fraud Layer)
imagehash         # Perceptual hashing for reverse-image check
requests          # SerpAPI / Winston AI HTTP calls

# ── Graph database ────────────────────────────────────────────────────────────
networkx          # In-memory Graph DB for Batch 2 multi-image fusion

# ── Database ──────────────────────────────────────────────────────────────────
pymongo           # MongoDB driver (primary DB backend)
dnspython         # Required for MongoDB Atlas SRV DNS lookups

# ── UI ────────────────────────────────────────────────────────────────────────
gradio>=4.44.0
pandas            # DataFrame support for financial tables
